webbackend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==4.2.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-ipware==6.0.3
django-rest-authemail==2.1.7
django-uuidfield==0.5.0
djangorestframework==3.14.0
git-filter-repo==2.38.0
idna==3.6
importlib-metadata==7.0.1
Markdown==3.5.2
mysqlclient==2.2.4
Pillow==9.0.0
pip==24.1.1
python-ipware==2.0.1
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
